src/robot/requirements.txt
 pyserial==3.5
opencv-python==4.8.1.78
numpy==1.24.3
openai==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
requests==2.31.0
python-dotenv==1.0.0